final touch

+++ frontend/src/components/Card.jsx
@@ -5,9 +5,12 @@ import EditForm from './EditForm';
 import {FaDumbbell} from 'react-icons/fa';
 import { IconContext } from "react-icons";
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 function Cards({ workout }) {
+
     const {dispatch} = useWorkoutsContext();
+    const {user} = useAuthContext();
     const [isModal, setIsModel] = useState(false)
     
     const editHandle = () =>{
@@ -18,8 +21,16 @@ function Cards({ workout }) {
     }
     
     const handleClick = async ()=>{
+        
+        if(!user){
+            return
+        }
+
        const response = await fetch(`http://localhost:4000/api/workout/${workout._id}`,{
-        method:'DELETE'
+        method:'DELETE',
+        headers:{
+            'Authorization': `Bearer ${user.token}`
+        }
        })
     const json = await response.json();
